refactor control
